Move small eventbuilding helpers into util.py
